+++ VE_Conductor/Framework/TexTable.cpp
@@ -85,8 +85,8 @@ void TexTable::SetNumofCols(int num)
 
 void TexTable::SetColTitles(const std::vector<wxString>& titles)
 {
-   for ( unsigned int i=0; 
-         (i < (size_t) titles.size() ) && ( i < (size_t) GetNumberCols() ); ++i)
+   for ( size_t i=0; 
+         (i < titles.size() ) && ( i < GetNumberCols() ); ++i)
    {
       SetColLabelValue( i, titles[i] );
    }
@@ -95,10 +95,10 @@ void TexTable::SetColTitles(const std::vector<wxString>& titles)
 void TexTable::SetColAlignments(const std::vector<int>& alignments)
 {
    m_align = alignments;
-   for ( unsigned int i=0; 
-         (i < (size_t) alignments.size() ) && ( i < (size_t) GetNumberCols() ); ++i )
+   for ( size_t i=0; 
+         (i <  alignments.size() ) && ( i < GetNumberCols() ); ++i )
    {
-      for ( unsigned int j = 0; j < (size_t) GetNumberRows(); ++j)
+      for ( size_t j = 0; j <  GetNumberRows(); ++j)
       {
          SetCellAlignment( j, i, alignments[i], wxALIGN_CENTER);
       }
